Update agents with core traits to enable versioned agents
