+++ crates/goose/src/agents/mod.rs
@@ -1,0 +1,7 @@
+mod agent;
+mod base;
+mod v1;
+
+pub use agent::Agent;
+pub use base::BaseAgent;
+pub use v1::AgentV1;
